ApiGateway - DomainName
ClusterStack name default to ecs-a

The idea is to do not change stack name between env.
StackName for resources, have an already long name (Ex. r-event-01), so
i use the same name for all env.
To have a consistent behaviour i will do the same for other stacktype
too.

LoadBalancerApplicationStack name default to alb-a

Same logic of ClusterStack name