* Trick to be able to add an element to a previuosly created/defined list
If the first element of a list is equals to "IBOXADDTOLIST", i try to
add the list to an already defined one (if exist).
Normalyy the new one overwrite the old.

This way i can centralize some general iampolicy in ecs-cluster
and attach them to RoleInstance, adding element to ManagedPolicyArns,
preserving the base ones.