AS_Autoscaling removed
moved Parameters and Outputs to autoscaling.yml

A key in role yaml can be enforced.

A key in role yaml can be enforced so that is value cannot be overriden
by env/region.
To enforce a key append '_IBOXENFORCE' to it's name.

Used on reserve-cpu to have the fixed value SkipClass for
CapacityDesired.