Use ImageIdLatest with value from SSM for ec2
This way there is no need to keep imageid-ec2.yml and imageid-ecs.yml
updated.
You can simply specify latest as ImageId value to use the latest
version available in the current region.

imageid-ec2.yml and imageid-ecs.yml remain as a reference for using a
specific ImageId in multiple used regions.