servicediscovery refactor
can be created in only one region.
Use parameter and mapping to enable it and to pass the id where the
resource is not created. Some method used for public hostedzones.

(Later i will do another refactor where i specify in a single
parameter/configuration which is the main region where to create the
global resources Ex. R53 zone, CF distro and so on ...)