Lambda ECSServiceAction renamed to LambdaECSEvents
Now the lambda hanlde ECS Task State Change events too.
So the name is now gneric for ECS Events.

Created a new events rule ECSTaskStateChange for triggering on
ECS TaskState Change.

The idea is to detect when a container is stopped for out of memory
error and notify with a sns publish on SNSTopicCloudWatchAlarm

renamed ECS events eventrules