WARNING big refactoring to S3 Replica!
Previuos commit about Replication are wrong.
I need to have multiple replica Prefix and for every Prefix i need to
create a ReplicationRule.
But every rules must be enabled and disabled based on an AWS If
Condition.

To achieve this i create a yaml cfg that reflect ReplicationRules
Structures and use auto_get_props.
But to create conditon automatically i have changed auto_get_props so
that is the key value is 'AWS::NoValue' i automatically create a
conditon and wrap value in it.

Work still in progress