WARNING VERY BIG CHANGE
LaunchTemplate is now near totally defined in
awsibox/cfg/BASE/launchtemplate.yml and value are taken using
auto_get_props.

I have changed the relative Parameters and Conditions used in
LaunchTemplate to reflect the new mapping names.

Work still in progress...

Important Note:

to accomplish this i had to find a way to wrapper returned obj in If
condition where needed.
So i introduced 3 new options:

To wrap an object recursed by auto_get_props add a property named: IBOXIF
IBOXIF must be a list where the first value is the condition and the
second the alternate value
Ex:
- AdditionalStorage:
    IBOXIF:
      - AdditionalStorage
      - Ref('AWS::NoValue')

To do the same in case of a string non recursed property create a string
space delimited where the first token is named IBOXIF, then put the name
of the condition followed by the value is case of true and the value in
case of false.
To wrap the property specify as value the name of the property itself.
Ex:
ImageId: IBOXIF LaunchTemplateDataImageIdLatest Ref('LaunchTemplateDataImageIdLatest') ImageId

This create the condition
If('LaunchTemplateDataImageIdLatest', Ref('LaunchTemplateDataImageIdLatest'), get_end_value('ImageId')