Fix over 18ece1b
key value to automatically create if condition and wrapper value in it
has been changed from AWS::NoValue to IBOXIFNOVALUE.
This has been done to better integrate in syntax used for if condition
of whole objects [IBOXIF].

Condition is now created so that value must differ from IBOXIFNOVALUE or
Ref('AWS::NoValue') is used (previous way was wrong and should not have
worked).

S3 Bucket specific changes:
AccessControlTranslation owner is fixed to Destination.
Is now created a "global" condition on whole AccessControlTranslation
key, using DestinationAccount as the "reference value" for it.
(AccessControlTranslation cannot be empty and must be created only if
DestinationAccount is used)

no need to use a try block for replace 'IBOXMAPNAME_'

bump version