clf_compute_order - fix
point subtracted for every char after a star need to weight very very
low. Otherway if there are a lot of char after a star priority can shift
in an unwanted way.
Ex consider:
/error-pages*/listaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa-*
/error-pages/*

The second one must have a precedence over the first, but if char weight
too much (Ex 1 point) this do not happen.
Star position differ only by one, this way the
presence of a great number of subsequent char overcame that difference
and change the resulting order.

To avoid this i subtract only 0.001 for every char after a star and the
half 0.0005, for a not first terminating star.