prepend_base_cfgs
Currenlty rds implementation to work with multiple DB defined in same
stack need to have DBInstance key named already configured in
awsibox/cfg/BASE/rds-dbinstance.yml; this is not the right way, as that
yaml conf file is integrated with awsibox and not exaternal.

So implemented a way to prepend yaml BASE configuration for an
object, using the already defined prepend_base_cfgs method.

If BASE_CFGS value for that key/object is IBOXBASE, when iterating over
cfg_cmm look far a subjey named Base, if found, for all other subkey of
the same object  prepend a configuration like the Base one.

Base changed to all uppercase BASE