taskdefinition to yaml
taskdefinition to yaml

_get_obj - do not look at obj_propname but use presence of key Value

_get_obj - work with class dict type too

New feature IBOXOBJOUTPUT

Used as IBXOBASE property in yaml parsed by auto_get_props,
same way as IBOXOUTPUT.
It's a way to defere the creation of an Output object if the Output
Value refer to a not already defined property of an object.

Lets' take ECSContainerDefinition:
We first use auto_get_props than we need to change some props value as
for example Cpu and Memory.
We want to create the ContainerDefinitions1Constraints Output in yaml
file but we cannot use normal IBOXOUTPUT, cause real values for Cpu and
Memory are defined later.
So the idea is:
- create an Output object as one created by IBOXOUTPUT using as value a
Sub like syntax
- alter it and add a property IBOX
- put as value of IBOX property the class of the object begin populated.
- then in stack_add_res, for outputs call a new method add_objoutput
that parse the Output sub value and replace the subkey with the
attribute of the object taken from IBOX property.

comment

_populate - key_value extend IBOXRESNAME match check to prefix

container definition to yaml (with great difficulty)

cloudfront fix

ECS_TaskDefinition - moved ContainerDefinition code in a  specific class