output listener
remove unused code

pep

Use Listener to popuate SecurityGroupLoadBalancer SecurityGroupIngress

The idea is:
for every Listeners i need to allow to connect to the relative of port
  of the loadbalancer.
So all information need to create the SG SecurityGroupIngresses can be
already present in the Listeners cfg/keys.

I changed SG_SecurityGroupRules, so that if SecurityGroupIngress is
empty i use cfg.Listeners to populate the SecurityGroupIngresses.
I had to make some changes to key prefix, condition and so on.

To avoid having to specify a SecurityGroupLoadBalancer with an empty
dict SecurityGroupIngress in every stack i added it to
loadbalancing-elb.yml, but i had to change ecs-alb.yml SecurityGroup
root_key using the double asterisk/star suffix **. This way the key
already present in SecurityGroup are removed (in this case the
SecurityGroupLoadBalancer key).

So at the end in Listeners i simply add the key Access that is later
"traslated" to  CidrIp.

fix to listerner output