new .IBOXCODE suffix to define the code wrapper of a key
Sometimes, like in autoscaling-scheduledactions.yml, a key value need to
be wrapped in an If Condition or some other CF code.
I need to have a base value that can be overriden later and a separate
code to describe how to wrapper and process that value.
I use a key named like {key}.IBOXCODE (same name of the value to
describe but with the suffix ".IBOXCODE").

In _populate method [awsibox/shared.py] i check for it and if found i
eval it's value.
I need to avoid overwring it's value, so for every propname i check if
exist a key named {propname}.IBOXCODE, in that case i skip processing
the propname key.