ECS Cluster Autoscaling
To properly work Cluster must be configured with a CapacityPrivider and
a defaultCapacityProviderStrategy.

The defaultCapacityProviderStrategy is used for a Service if no other
CapacityProviderStrategy is specified BUT ONLY if the LaunchType
property is NOT provided!
(discovered this after many hours of unsuccessefully experiments)

So i need to make some changes in particular to be able to update an
existing Cluster with already created services.

*) Need to add CP ManagedScaling Status property.
This way i can add the CP to the Cluster but having ManagedScaling
DISABLED so that ASG is not affected.
As i first need to update services removing LaunchType property.

*) For ECS Services i need to remove the LaunchType property if
LaunchType != FARGATE. This way they will use the provided
defaultCapacityProviderStrategy.

So the workflow for an existing Cluser and Services is.
*) Create and associate CP to Cluster with ManagedScalingStatus =
DISABLED
*) Update services removing LaunchType.
*) Activate ManagedScaling for the CP