BEGIN big change - Use True [true] and False [false] in place of None for condition
Should be more clear to only use True and False (values in template "true" and "false")
in condition replacing "None".
I begin with scheduled action recurrence and schedule, cause at the
moment it's not used (and i need to use it).
But i will need to process all py and yaml files looking for None.

Add support for MaxInstanceLifetime

Use IBOXPCO for Parameters and ouput, it must be inserted before
correspective key name, or it will not trigger the automatic override on
parameter value.
Use IBOXTITLE to replace the title after having processed object with
auto_get_props