* new features: automatically create parameter and ouput
If for a final property key named {propname} there is a key named like:
{propname}.IBOX_AUTO_PO process it to automatically create parameter and
output.
(For example look in cfg/BASE/ecs/ecs-service.yml [PlacementStrategies])

The default Value are:
for parameter:
  'Description': 'Empty for mapped value'
for output:
  Value: {the relative final value of the processed key/parameter}

To use them the {propname}.IBOX_AUTO_PO key value must be an
empty dict "{}"

If you want to change some values or add other (Ex AllowedValues) you
can create them as dict key/value.
(The default one will be updated with the specified ones)
Ex:
  Type.IBOX_AUTO_PO:
    AllowedValues: ['', 'binpack', 'random', 'spread']

I had to split parameter and output creation in two phase:

- Parameter need to be created before looking for key/property value (or
override condition will not be created and taken in account).

- Output need to be created after discovering final property value as
usually it's value is the same, this way any condition present in the
value will be automatically present in output too.