Cfg root nodes for EnvRole and EnvType (TYPE.yml)
Are no more named as EnvRole or TYPE file but are always named "global"
as normal root node of included files.

This allow to include in a file of type ECS the configuration of another
one (Ex ecs-cluster).

I am experimenting to create a configuration that permit to have a stack
that create an ECS service and at the same time a dedicated "personal" ECS Cluster.