New condition for containers: UseTaskCpu
If containers Cpu value equals "use_task_cpu" and if taskdefinition Cpu
is defined (not none), containers Cpu value will equals the one of the
task.

First this was enforced if there was only one container definition.
But it can be usefull if running a task on FARGATE to use a container
cpu value less than the one used for the task.
The container cpu value is used for cloudwatch metric so is used by the
targettracking policies too.
Container cpu is a soft limit.
Ex:
Task Cpu 2048 (2 core)
Container Cpu: 1024

container can access full 2 core cpu, but metric will be relative to
only 1 core (100% = 1 core).