begin support for networkloadbalancer
support for networkloadbalancer - progress

support for networkloadbalancer - progress

ListenerLoadBalancerHttp/sPort parameter and condition

they are created ONLY for the relative listener.
This is more logic, if port is 80 (http) or 443 (https) i do not create
the listener at all during template creation.
So it make no sense to have parameter for changing port, wihout lister
even if we change port it do not have effect.
Otherway if i already create listerner (http and https port differ from
80 and 443) i create the parameter, condition (maybe only http or https)
and if i change the port, i really change the listener.

(network LB progress too)

yaml fix

** BIG CHANGES - ListenerLoadBalancerHttp/sPort **

To simply template, remove un-used or not so usefull condition and
parameters.

For CloudFront distro we simply configure origin HTTP/S port looking at
ListenerLoadBalancerHttp/sPort; no more condition on
OriginProtocolPolicy.
As a consequence ListenerLoadBalancerHttp/sPort cannot assume no more
value 'none' so relative condition is removed.
So we remove all Resource Condition on Listener and ListenerRules.
Usually the default rule of:
"create http rules on Internal LB and https rules on External one"
is sufficient.
It's not a problem if we create an https listener on the External LB
even if CloudFront is configured to connect using the http.
And there is already code to enforce a different behaviour.

cosmetic