build_RP - change to make ** override recurse
previously ** method to override a previous key value (if a dict) was
working only for first level key (Ex SecurityGroup in ecs-alb).
Now i remove the suffix ** only if there is an exiting key to override,
otherway i leave it; this way it can be later intercepted by merge
method that recurse while merging dict.
This way ** method can be used in subkeys too (Ex PortMappings)