cfg - dev/stg/prd conf
Do not use a structure (dict) for environment/region configuration.
This way is more clear that it's used only for the automatic creation
of the Mapping section and is not used for the structure/resources.

As advantage configuration keys name for environment/region can be "splitted"
as desired, Ex:

ScheduledAction:
  DownStartTime: 2100-01-01

can be written:

ScheduledAction:
  Down:
    StartTime: 2100-01-01

or even on a single line:
ScheduledActionDownStartTime: 2100-01-01