LambdaFunctionAssociation - fix to use Condition
LambdaFunctionAssociation need to warapped in If Condition.
To do this i add a Condition property in yaml cfg to the
LambdaFunctionAssociation.
Later in `add_obj` i check if the obj is a Resource property (and not a
Resource itself) where Condition is not supported.
In that case i remove the Condition and wrap the obj in an If Condition.