cosmetic - before change
ECSDrainInstance - registered/remainingResources

Current method is not usable for a large number of tasks running on
container instance.

Main problem that take me to the first ECSDrainInstance refactor is that
seems that when a task is draining "runningTasksCount" report 0 even if
the task is still running waiting for LB/TargetGroup draining.
So first idea was to keep track of the task in running state just after
draining the instance and then look for the presence of the filed stoppedAt in the
decsribe_tasks reuslts.

But if there are a lot of task, this can take time.

With this commit i try to use the field remainingResources and
registeredResources for the CPU resource, hoping that they are not
involved by the same problem.