ECSDrainInstance, retry in draining for INVALID_INSTANCE_STATE_TRANSITION
Usually if container instance change state to DRAINING return error with
reason "INVALID_INSTANCE_STATE_TRANSITION" means that instance is still
registering (is not already ACTIVE).
So if we re-try as usual sending the msg after 40sec, in the meanwhile
tasks can be started on the ci that need to be drained.
Better to retry for n=5 times sleeping 5 sec between retries, until
instance register so we can successfully DRAIN it.

(shorten some log message to reduce code size under 4096)