parse_ibox_key, use it for IBOX_TITLE too
parse_ibox_key, litte refactor

to avoid confusionits' better to use a different notation
and not the same of f-string.

Simply define a tuple to identify IBOX special keys and in turn try to replace
them with the defined value.
A dot "." can be used to write a key using with IBOX key and string in a
more readable value.
(it's simply removed by parse_ibox_key during parsing)

Ex:
IBOX_MAPNAME.Account