linked_obj progress
end work on IBOX_LINKED_OBJ

For cycle is always executed, even if missing because default to [""].
If missing index key will be empty ("") so it will not change the name.

If _source_ obj and _target_ are the same i update the existing object
other way i create a new one based on the source.

Look as docs, IBOX_SPECIAL_KEYS.md,  for usage examples.