all lambda to joker, progress
IBOX_LINKED_OBJ can be a dict

It's better over a list, because this way i can update/overwrite a configuration
later using the index name.
But i have the problem to identify when is a single conf or a dict of
conf as the single conf is a dict too.
I check for the not presence of a key named "Type" always found inside
the "base" conf.
It's a bad hack, but at the moment i have no other ideas.