IBOX_PCO for Custom Key ONLY
If there is an IBOX_PCO key for a custom props, a props not of the aws
resource obj, process it but only if there is not a relative IBOX_AUTO_P/O
or IBOX_CODE key.

This way i can move Bucket Output to cfg, and use the presence of the
OutputValueRegion key to override an already defined IBOX_OUTPUT with a
different one.