Support for skipping mappedvalues for specif key
If a final key, the one with the final/end value endswith "@", ignore
any mapped values if exists and use it's value for all env/region.

Usefull if i have a key value that change for TYPE and i want to have it
fixed for EnvRole without override it in all region env where it have
been previously defined.