Conditionally create LogGroup for Lambda
at the end i prefer to have a Template Condition with a Parameter too,
this way i can disable/enable it specifically for the single stack.
Ex i have a stack in eu-west-1 and want to create the same in a new region,
i would like that the new one create the LogGroup too.
This way i can simply disable by Parameter the LogGroup creation for the
current already in use stacks.