IBOX_AUTO_PO can be used on custom key too
IBOX_AUTO_PO can now be used also on custom key (the one not used by aws
resources), but only if:
the relative key (propname) is not present in the conf.
As example:
i can use testKey.IBOX_AUTO_PO but must not exist a testKey key.

It's automatically created a parameter, a condition if the Condition sub
key does exist and an output with as value the value of the parameter
(Ref).

use new custom IBOX_AUTO_PO on ImageIdLatest