refactor cloudfront/for-services CloudFrontDistribution
Removed condition RecordSetCloudFront, it make no sense to use
Ref('RecordSetExternal') as Alias of the CloudFrontDistribution!

USed IBOX_LINKED_OBJ_NAME to create RecordSetCloudFront, this is the way
to maybe allow in the future multiple Distribution for Stack.
At the moment the ResourceTitle is fixed to RecordSetCloudFront.

RecordSetCloudFrontSuffix have been replaced with a more appropriate
one, directly inside the Resource:
DnsSuffix

Ex: CloudFrontDistributionBaseDnsSuffix