ECSEventServiceFailure
create ad hoc inline lambda for ECSEventServiceFailure in case of ECS
Service is unable to place a task for lack of cpu or memory.
This trigger a lambda that raise a custom ad hoc alarm to increase the
relative cluster ASG size.

I have splitted the events from event_ecs-spot.yml, but this event and
relative lambda is not used anymore and is not included.
AWS Cluster Autoscaling do all this in a better way, so i simply retain
it as possible source for other idea and implementation.