ASG InstanceMaintenancePolicy
If on ASG is enabled MaxInstanceLifetime and
NewInstancesProtectedFromScaleIn instaces are not terminated after
MaxInstanceLifetime period because usually termination event is
triggered before launching a new istance to replace the terminating one.

USing InstanceMaintenancePolicy with MaxHealthyPercentage = 200 and
MinHealthyPercentage 100 should ensure that a new instance is launched
before terminating the old one.