lambdas/ECSClusterAutoscale, consider AZ too
when terminating instance take in account AZ too.
Primary condition is to terminate instance in the AZ with the biggest
number of instances running.
Then look at number of tasks running and registeredAt (prefer the
oldest one).
This way i avoid to create inbalance in AZ that are later fixed by AWS
by terminating and launching new instance.