ASG InstanceMaintenancePolicy, fox for -1 value
Value -1 cannot be used on creation, but only to remove a previously
setted value.
And i cannot specity -1 only for min or max but for both at the same
time.
This way i prefer to define another key:
InstanceMaintenancePolicyEnabled

and add condition on that key.

By default policy is disabled, if enabled default values are:
MaxHealthyPercentage: 200
MinHealthyPercentage: 100