lambdas/CCRAWSResource.code, bad fix for:
aws-cloudformation issues/1037
The properties are always returning String even though they have been passed as String, Boolean and Integer.

This do not resolve the problem, cause could be a string that can be
converted to int but that the service want as string.
For the moment i keep it this way, maybe (but i do not think) will take
the effort to fix this???????