BIG REFACTOR begin, do not use anymore global trick in shared
to be able to use IBOX keys values as IBOX_RESNAME inside eval, until
now i had to define them as globals of shared.py, but there was the problem that
during mutiple template creation shared globals where not resetted and
sometimes there is a race condition and some value of the previous run "
contaminate" the subsequent one.
I could reset them, but i prefer a more robust approach.
I use a dict class changed to support inserting and querying values as
attr: cfg.BUILD_ENVS
The dict object is created in the initialization script.

To simplify the call to eval i created and helper function ibox_eval,
that always use as globals the dict cfg.BUILD_ENVS.

fix and progress