get_condition, parse value2 too
until now value2 in get_condition was treated as a fixed value, string
or int.
Now it can be the name of a key to be parsed.
This way should be possible to automatically create a condition that
test the value of two different keys that can be overriden by parameters